pyproject.toml
[project]
name = "sigmak"
version = "0.1.0"
description = "Proprietary Risk Scoring API for SEC 10-K/Q filings"
readme = "README.md"
authors = [
    { name = "Peter Martinson", email = "peter.j.martinson@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "requests>=2.31.0",
    "chromadb>=0.4.0",
    "pytest>=8.0.0",
    "langchain-text-splitters>=1.1.0",
    "sentence-transformers>=5.2.0",
    "numpy>=2.2.6",
    "scikit-learn>=1.7.2",
    "bs4>=0.0.2",
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "python-multipart>=0.0.21",
    "fastapi-security>=0.1.0",
    "slowapi>=0.1.9",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "psutil>=5.9.0",
    "google-genai>=0.2.0",
    "apscheduler>=3.10.0",
    "pyyaml>=6.0.3",
    "yfinance>=1.1.0",
    "vaderSentiment>=3.3.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
pdf = [
    "weasyprint>=62.0",
    "markdown>=3.5.0",
    "jinja2>=3.1.0",
]
edgar = [
    "edgartools>=2.0.0",
]

[dependency-groups]
dev = [
    "flake8>=7.0.0",
    "mypy>=1.9.0",
    "types-requests",
    "types-beautifulsoup4",
]

[tool.mypy]
python_version = "3.11"

# --- Global Strictness Settings ---
disallow_untyped_defs = true    # Force type hints on all functions
check_untyped_defs = true       # Check bodies of untyped functions
no_implicit_optional = true     # Force explicit Optional[X] or X | None
warn_return_any = true          # Warn when returning a value of type Any
warn_unused_configs = true      # Warn about unused config sections
warn_unused_ignores = true      # Warn about unnecessary '# type: ignore' comments

# --- Third-Party Library Overrides ---
# These libraries either don't have type stubs or have incomplete ones.
# We silence 'missing import' errors specifically for these modules.
[[tool.mypy.overrides]]
module = [
    "chromadb.*",
    "bs4.*",
    "langchain_text_splitters.*",
    "lxml.*",
    "google.genai.*",
    "vaderSentiment.*",
    "edgar.*"
]
ignore_missing_imports = true
